Add Autocomplete utils file. Move menu keyboard navigation to useKeyboardNavigation
